Promoted fluent to directory, refactor to reduce code duplication
